add turn geometry column to list.html
minor update to how we check if the current page list count is greater than zero. also, there is a commented column for checking the static value of speed rating against our computed. this can be helpful for now while we still have to maintain the static speed value for sorting